Statistics calculates variance
* DecisionMaker uses variance to calculate the amount of asset that should be bought.

* Add comments to some packages functions;

* Statistics code clean up.
